Wait for coral to be completely shot through funnel before reseting elevator.
After shooting coral, we set the elevator to the bottom hardstop, which actually sets the position an inch above the hardstop and then lowers and stalls the elevator and resets the position to 0. If we don't wait for the coral to be completely shot through the funnel, we raise the elevator to the 1" position while the coral is still exiting the manipulator.