Post NLR Vision Improvements (#46)
* Fix vision target logging to only include those with accepted poses.

Previously, we logged all vision targets that provided poses; even those whose poses were rejected. This will make it more clear which vision targets are providing accepted poses when viewing log files.

* Only accept vision pose estimates if the estimated pose's rotation is close to the robot's current rotation.

Sometimes vision provides a pose estimate with a reasonable ambiguity but it clearly the wrong reprojection. When this happens, the rotation of the pose estimate is usually quite different than the robot's current rotation.

* Log the robot-to-camera transform with a known robot pose.

* Add camera transforms tab.

* Consistently restrict cameras to consider when executing drive-to-pose.

* Tune max z error to avoid bad pose estimates.

* Update funnel supply current.